backend/requirements.txt
fastapi
uvicorn
geopandas
rasterio
google-generativeai
python-dotenv
python-multipart
shapely
requests
numpy
ultralytics
opencv-python
pillow
pyproj
